fix(j-s): Renaming
